WIP auth key middleware
