--- packages/redwoodjs/console
@@ -1,2 +1,0 @@
-{"level":60,"time":1715026201796,"pid":94751,"hostname":"depthcharge-air-3303.local","msg":">>> Unkey bopom"}
-{"level":50,"time":1715026201796,"pid":94751,"hostname":"depthcharge-air-3303.local","msg":"Error in withUnkey"}
